pyproject.toml
[project]
name = "projeto-streaming-gcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-pubsub>=2.28.0",
    "python-dotenv>=1.0.1",
]
